fix: redirect unauthenticated on admin routes
