+++ src/spa/router/guards/AuthenticatedRouteGuard.tsx
@@ -1,24 +1,15 @@
-import { FC, useEffect } from 'react';
-
-import { useLocation, useNavigate } from 'react-router-dom';
+import { FC } from 'react';
 
 import { ErrorBoundary } from '@/components/ErrorBoundary';
 import { useAuthContext } from '@/spa/auth/AuthContext';
+import { useRedirectUnauthenticated } from '@/spa/router/guards/useRedirectUnauthenticated';
 
 export const AuthenticatedRouteGuard: FC<React.PropsWithChildren<unknown>> = ({
   children,
 }) => {
   const { isAuthenticated } = useAuthContext();
-  const { pathname, search } = useLocation();
-  const navigate = useNavigate();
 
-  useEffect(() => {
-    if (!isAuthenticated) {
-      navigate(`/login?redirect=${encodeURIComponent(pathname + search)}`, {
-        replace: true,
-      });
-    }
-  }, [isAuthenticated, navigate, pathname, search]);
+  useRedirectUnauthenticated();
 
   return !isAuthenticated ? null : <ErrorBoundary>{children}</ErrorBoundary>;
 };
